fix thu

+++ api/controllers/CommunicationController.js
@@ -12,25 +12,68 @@ module.exports = {
     res.header('Access-Control-Expose-Headers', 'Authorization')
 
     var token = req.header('Authorization');
-    console.log(token);
 
-    let executeQuery = function (res, query) {
-      new sql.ConnectionPool(config.dbConfigSetting).connect().then((pool) => {
-        return pool.request().query(query);
-      }).then((data) => {
-        let reult = {
-          communications: data.recordsets[0]
-        }
-        sql.close();
-        return res.status(200).json(reult);
-      }).catch((err) => {
-        sql.close();
-        return res.status(500).send({ message: `${err}` });
-      });
-    }
-    let query = "SELECT * FROM communication";
-    executeQuery(res, query);
+
+
+
+    if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+    jwt.verify(token, 'supersecret', function (err, decoded) {
+      if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+
+      let executeQuery = function (res, query) {
+        new sql.ConnectionPool(config.dbConfigSetting).connect().then((pool) => {
+          return pool.request().query(query);
+        }).then((data) => {
+          let reult = {
+            communications: data.recordsets[0]
+          }
+          sql.close();
+          return res.status(200).json(reult);
+        }).catch((err) => {
+          sql.close();
+          return res.status(500).send({ message: `${err}` });
+        });
+      }
+      let query = "SELECT * FROM communication";
+      executeQuery(res, query);
+
+    })
+
+
   },
+
+
+
+  getNotifyAllCommunication: async (req, res) => {
+    res.header('Access-Control-Allow-Origin', '*');
+    res.header('Access-Control-Allow-Methods', 'GET, POST, PATCH, PUT, DELETE, OPTIONS');
+    res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept,Content-Type, Authorization');
+    res.header('Access-Control-Allow-Credentials', true);
+    res.header('Access-Control-Expose-Headers', 'Authorization')
+
+    var token = req.header('Authorization');
+
+      let executeQuery = function (res, query) {
+        new sql.ConnectionPool(config.dbConfigSetting).connect().then((pool) => {
+          return pool.request().query(query);
+        }).then((data) => {
+          let reult = {
+            communications: data.recordsets[0]
+          }
+          sql.close();
+          return res.status(200).json(reult);
+        }).catch((err) => {
+          sql.close();
+          return res.status(500).send({ message: `${err}` });
+        });
+      }
+      let query = "SELECT * FROM communication";
+      executeQuery(res, query);
+
+  },
+
+
 
   postCommunication: async (req, res) => {
 
@@ -41,9 +84,20 @@ module.exports = {
     // res.header('Access-Control-Expose-Headers', 'Authorization')
 
     var token = req.header('Authorization');
-    console.log(token);
+    
+    
 
     let newCommunication = req.body;
+
+
+
+    if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+    jwt.verify(token, 'supersecret', function (err, decoded) {
+      if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+
+
+
     let executeQuery = function (res, query) {
       sql.connect(config.dbConfigSetting, function (err) {
         if (err) {
@@ -71,6 +125,7 @@ module.exports = {
         values('${newCommunication.image}',   N'${newCommunication.title}', N'${newCommunication.content}', N'${newCommunication.action_discount}', '${newCommunication.link_communication}', '${newCommunication.category}', '${newCommunication.is_active}', '${newCommunication.locations}', '${newCommunication.sortId}')
         SELECT * FROM communication WHERE id = SCOPE_IDENTITY()`;
     executeQuery(res, query);
+  })
 
   },
 
@@ -80,10 +135,18 @@ module.exports = {
     res.header('Access-Control-Allow-Methods', 'GET, POST, PATCH, PUT, DELETE, OPTIONS');
 
     var token = req.header('Authorization');
-    console.log(token);
+
 
 
     let id = req.param('id');
+
+
+    if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+    jwt.verify(token, 'supersecret', function (err, decoded) {
+      if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+
+
     let executeQuery = function (res, query) {
       sql.connect(config.dbConfigSetting, function (err) {
         if (err) {
@@ -107,6 +170,7 @@ module.exports = {
     }
     let query = `EXEC UpdateNotify ${id}`;
     executeQuery(res, query);
+  })
   },
 
   sortCommunication: async (req, res) => {
@@ -116,7 +180,13 @@ module.exports = {
     res.header('Access-Control-Allow-Methods', 'GET, POST, PATCH, PUT, DELETE, OPTIONS');
 
     var token = req.header('Authorization');
-    console.log(token);
+
+
+
+    if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+    jwt.verify(token, 'supersecret', function (err, decoded) {
+      if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
 
 
 
@@ -136,12 +206,14 @@ module.exports = {
     let objSortId = req.body;
     let arrSortId = objSortId.updateSordId;
 
-    if(arrSortId){
+    if (arrSortId) {
       arrSortId.forEach(element => {
         let query = `EXEC SortNotify ${element.id}, ${element.sortId}`;
         executeQuery(res, query);
       });
+      
     }
+  })
 
   },
 
@@ -152,31 +224,39 @@ module.exports = {
 
 
     var token = req.header('Authorization');
-    console.log(token);
 
-    
+
+
 
     let id = req.param('id');
-    let executeQuery = function (res, query) {
-      sql.connect(config.dbConfigSetting, function (err) {
-        if (err) {
-          console.log("Error while connecting database :- " + err);
-          res.send(err);
-        } else {
-          var request = new sql.Request();
-          request.query(query, function (err, data) {
-            if (err) {
-              console.log("Error while querying database :- " + err);
-              res.send(err);
-            } else {
-              sql.close();
-              return res.json({ status: " Status 200 ok" });
-            }
-          });
-        }
-      });
-    }
-    let query = `DELETE FROM communication WHERE id= ${id}`;
-    executeQuery(res, query);
+
+    
+    if (!token) return res.status(401).send({ auth: false, message: 'No token provided.' });
+
+    jwt.verify(token, 'supersecret', function (err, decoded) {
+      if (err) return res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+      let executeQuery = function (res, query) {
+        sql.connect(config.dbConfigSetting, function (err) {
+          if (err) {
+            console.log("Error while connecting database :- " + err);
+            res.send(err);
+          } else {
+            var request = new sql.Request();
+            request.query(query, function (err, data) {
+              if (err) {
+                console.log("Error while querying database :- " + err);
+                res.send(err);
+              } else {
+                sql.close();
+                return res.json({ status: " Status 200 ok" });
+              }
+            });
+          }
+        });
+      }
+      let query = `DELETE FROM communication WHERE id= ${id}`;
+      executeQuery(res, query);
+    })
+    
   }
 };
